chore: note page updates
