+++ src/components/modules/Note/NoteCardContent.astro
@@ -5,7 +5,20 @@ interface Props {
 }
 const { pageId } = Astro.props;
 const apiKey = import.meta.env.NOTION_API_KEY;
-const pageContent = await getPageContent(pageId, { apiKey });
+const pageContent = await getPageContent(pageId, {
+  apiKey,
+  accessControl: (resp) => {
+    const isValid =
+      "created_time" in resp &&
+      resp.parent.type === "database_id" &&
+      resp.parent.database_id.replace(/-/g, "") ===
+        import.meta.env.NOTION_NOTE_DATABASE_ID &&
+      resp.properties["Is Published"].type === "checkbox" &&
+      resp.properties["Is Published"].checkbox === true;
+
+    return isValid;
+  },
+});
 ---
 
 <Fragment set:html={pageContent} />
