+++ src/libs/notion-client/index.ts
@@ -1,2 +1,2 @@
 export * from "./getEntries";
-export * from "./getBlock";
+export * from "./getPage";
